Refactor to split server and client
